Add users, delete users, updated quantity on subscription.

+++ resources/views/dashboard/company/index.blade.php
@@ -46,21 +46,23 @@
 					                            <br />
 					                         
 					                        <!-- Split button -->
-					                        <div class="btn-group">
-					                            <button type="button" class="btn btn-primary btn-sm">
-					                                Actions</button>
-					                            <button type="button" class="btn btn-primary btn-sm dropdown-toggle" data-toggle="dropdown">
-					                                <span class="caret"></span><span class="sr-only">Social</span>
-					                            </button>
-					                            <ul class="dropdown-menu" role="menu" data-form="deleteForm">
-					                                <li><a href="{{ route('users.create')}}">Add Users</a></li>
-					                                <li><a href="{{ route('company.edit', ['id' => $company->id]) }}">Edit Company</a></li>
-						                            <li><a href="{{ route('company.destroy', $company->id) }}" class="form-delete" data-method="delete" name="delete_item" data-token="{{ csrf_token() }}">Delete Company</a>
-						                            </li>
-					                                <li class="divider"></li>
-					                                <li><a href="{{ url('logout') }}">Sign Out</a></li>
-					                            </ul>
-					                        </div>
+					                        @if ( Auth::user()->hasRole('super_admin') )
+						                        <div class="btn-group">
+						                            <button type="button" class="btn btn-primary btn-sm">
+						                                Actions</button>
+						                            <button type="button" class="btn btn-primary btn-sm dropdown-toggle" data-toggle="dropdown">
+						                                <span class="caret"></span><span class="sr-only">Social</span>
+						                            </button>
+						                            <ul class="dropdown-menu" role="menu" data-form="deleteForm">
+						                                <li><a href="{{ route('users.create')}}">Add Users</a></li>
+						                                <li><a href="{{ route('company.edit', ['id' => $company->id]) }}">Edit Company</a></li>
+							                            <li><a href="{{ route('company.destroy', $company->id) }}" class="form-delete" data-method="delete" name="delete_item" data-token="{{ csrf_token() }}">Delete Company</a>
+							                            </li>
+						                                <li class="divider"></li>
+						                                <li><a href="{{ url('logout') }}">Sign Out</a></li>
+						                            </ul>
+						                        </div>
+					                        @endif
 					                    </div>
 					                </div>
 					            </div>
